dcp-815 separate protocol_link.py
